pyproject.toml
[project]
name = "poly2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9.10"
dependencies = [
    "dataclasses>=0.8",
    "matplotlib>=3.9.4",
    "pandas>=2.3.3",
    "py-clob-client>=0.34.5",
    "pyarrow>=21.0.0",
    "requests>=2.32.5",
    "websocket-client>=1.9.0",
    "websockets>=15.0.1",
]
